Move eq_op::is_exp_equal to utils
